Added new arguments for isdb-tb
The new options forces the setup of isdb-tb tuners, allowing wscan2 to generate a channel list for this standard (tested with MyGica S2870 & PlayTV USB Hybrid PV-B308U)

BTW for now the channel width selection is forced to 6Mhz, but wikipedia states that there are countries using 8Mhz (e.g. Maldives) so it may be necessary to do some work on that matter.